refactor: modify JPARepository to Repository
