--- backend/src/main/java/com/chillin/hearting/db/repository/NotificationJPARepository.java
@@ -1,12 +1,0 @@
-package com.chillin.hearting.db.repository;
-
-import com.chillin.hearting.db.domain.Notification;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface NotificationJPARepository extends JpaRepository<Notification, Long> {
-
-    List<Notification> findByUserIdAndIsActiveTrue(String userId, Sort sort);
-}
